UnlockTech works
New command, unlock tech, allows unlocking of a tech associated with a techguid

+++ APVRising/Hooks/ChatMessage.cs
@@ -56,6 +56,24 @@ public static class ChatMessage
         users.Dispose();
     }
 
+    public static void NotifyClientUnlock(int guid)
+    {
+        // Find the chat message system
+        var em = Plugin.EntityManager;
+        var userQuery = em.CreateEntityQuery(ComponentType.ReadOnly<User>());
+        var users = userQuery.ToEntityArray(Allocator.Temp);
+        Plugin.BepinLogger.LogInfo($"Notifying clients of AP list change: GUID={guid}, Users={users.Length}");
+        foreach (var userEntity in users)
+        {
+            var user = em.GetComponentData<User>(userEntity);
+            // Send a special prefixed message the client can intercept
+            var message = (FixedString512Bytes)$"##AP_UNLOCK#{guid}##";
+            ServerChatUtils.SendSystemMessageToClient(em, user, ref message);
+        }
+
+        users.Dispose();
+    }
+
     [HarmonyPatch(typeof(ClientChatSystem), "OnUpdate")]
     [HarmonyPrefix]
     public static void ClientChatOnUpdatePostfix(ClientChatSystem __instance)
@@ -76,8 +94,31 @@ public static class ChatMessage
                 {
                     bool isResearching = message.Contains("True");
                     ProgressionHandler.IsResearching = isResearching;
+                    ProgressionHandler.isStale = true;
+
                     Plugin.BepinLogger.LogInfo($"Client AP state: IsResearching={isResearching}");
 
+                    // Destroy so it doesn't appear in chat UI
+                    em.DestroyEntity(eventEntity);
+                }
+                if (message.StartsWith("##AP_UNLOCK#"))
+                {
+                    string guidStr = message.Replace("##AP_UNLOCK#", "").Replace("##", "");
+                    if (int.TryParse(guidStr, out int guid))
+                    {
+                        Plugin.BepinLogger.LogInfo($"Client AP unlock: GUID={guid}");
+                        ArchipelagoData.APProgression.Add(guid);
+                        var userQuery = em.CreateEntityQuery(ComponentType.ReadOnly<User>(), ComponentType.ReadOnly<ProgressionMapper>());
+                        var userEntities = userQuery.ToEntityArray(Allocator.Temp);
+                        foreach (var userEntity in userEntities)
+                        {
+                            ProgressionHandler.UnlockResearchForPlayer(userEntity, new Stunlock.Core.PrefabGUID(guid));
+                        }
+                    }
+                    else
+                    {
+                        Plugin.BepinLogger.LogError($"Failed to parse AP unlock GUID from message: {message}");
+                    }
                     // Destroy so it doesn't appear in chat UI
                     em.DestroyEntity(eventEntity);
                 }
